refactor: move createEventFromProposalAction to proposal-action-events.service
