+++ src/events/events.service.ts
@@ -15,7 +15,6 @@ import { GroupPrivacy } from "../groups/group-configs/models/group-config.model"
 import { randomDefaultImagePath, saveImage } from "../images/image.utils";
 import { ImagesService, ImageTypes } from "../images/images.service";
 import { Image } from "../images/models/image.model";
-import { ProposalActionEvent } from "../proposals/proposal-actions/proposal-action-events/models/proposal-action-event.model";
 import { EventAttendeesService } from "./event-attendees/event-attendees.service";
 import {
   EventAttendee,
@@ -190,25 +189,6 @@ export class EventsService {
       await this.saveDefaultCoverPhoto(event.id);
     }
     return { event };
-  }
-
-  async createEventFromProposal(
-    { images, ...eventData }: ProposalActionEvent,
-    hostId: number
-  ) {
-    const event = await this.eventRepository.save(eventData);
-    const coverPhoto = images.find(
-      ({ imageType }) => imageType === ImageTypes.CoverPhoto
-    );
-    await this.eventAttendeeRepository.save({
-      status: EventAttendeeStatus.Host,
-      eventId: event.id,
-      hostId,
-    });
-    await this.imagesService.createImage({
-      ...coverPhoto,
-      eventId: event.id,
-    });
   }
 
   async updateEvent({ id, coverPhoto, ...eventData }: UpdateEventInput) {
